Make pagination more resiliant.
Browse URL now takes the page number in the current order as it's first
argument rather than an index into the event history. The argument can
also be an event ID in which case the page containing that event is
shown.

Event ID generation has been changed so all IDs begin with an 'e'.
Otherwise they may parse as integers in the browse URL.

When viewing events from the browse page the count and ascending values
are passed to the event so the browse link in the event can return to
the same page with the same values.