Fix bug where series banners may not show in feed.
When events are added we do not wait for the ensureSeries call to
complete. This caused problems if the event was then added to the feed
as the series data may not yet have been fetched causing banner images
not to be shown in the feed.

We now also now call ensureSeries when adding to the feed. The function
was also improved to not refetch series data if call for the series was
already inflight.