Improve resiliance to Sonarr connection issues.
If Sonarr is not connectable at startup, or becomes unconnectable, then
Sonarr to RSS was unable to reconnect to fetch series data and banner
images once Sonarr became available again.

This situation could easily be encountered in Docker setups where both
Sonarr and Sonarr to RSS are running on the same machine. Sonarr to RSS
will startup much faster than Sonarr so if both containers are
automatically started on system boot Sonarr will not be connectable and
the api never becomes available.

All code paths that make calls to Sonarr now make an ensure call to
re-initialise the api if it is unavailable. They also clean up correctly
on failure so that future calls will be remade.

Fixed typo and name styling in README.md.